Standardize on typedef struct foo_bar_st FOO_BAR.
Some of the struct names are very different from the typedef. Fix these and fix
a few references to refer to the typedef versions. Make this consistent so it
is easier to grep for struct definitions.

The ones that are off only in capitalization are left alone. I found a handful
of consumers with forward declarations that would break if they were
lowercased, and having to grep -i is at least not quite as absurd as
remembering

  typedef struct env_md_ctx_st EVP_MD_CTX;

Change-Id: Ie6e69146c650836bc661601c6e9e765f07e114dd