Update patch set 1
Patch Set 1:

(2 comments)

Rebased and updated to document the maximum for all functions since it's also true for EVP_PKEY_sign for ECDSA keys. It's not actually true for EVP_PKEY_encrypt and I don't know about EVP_PKEY_derive because nothing implements it. But it seems cleaner to just not promise it for any at that point.

Patch-set: 1
Label: Code-Review=0