Update patch set 2
Patch Set 2: Code-Review+2

LGTM.

But note the comment at the top of the function - I'm pretty sure that padding check functions need to be constant time, but many of these aren't. I think many of the check functions need to be rewritten, which is why I didn't bother too closely with the current ones.

Patch-set: 2
Label: Code-Review=+2