Consolidate CCS_OK paths in s3_clnt.c.
Introduce a CR_CHANGE state just before entering CR_FINISHED_A along with a
CR_CHANGE_FLUSH state to combine preparing for ChangeCipherSpec and flushing
the write buffer.

If this is a renegotiation, the application data protocol is still running and
adding data to the write buffer. Thus BIO_flush may block until more data than
necessary is written out to network. While it is blocking unnecessarily, the
peer has received the Finished and may send ChangeCipherSpec. The application
data protocol is running at this point, so it may read records and process
ChangeCipherSpec.

I believe this to be the cause of
https://rt.openssl.org/Ticket/Display.html?id=3400 and why it was necessary to
add a CCS_OK in s3_clnt.c. This is worth double-checking, however.

Change-Id: I6a928386994fcd5efff26a5f0efb12b65bf7f299