Consolidate CCS_OK codepaths in s3_srvr.c.
Rename SSL3_ST_SR_POST_CLIENT_CERT to SSL3_ST_SR_CHANGE and have this be the
point at which CCS_OK is set. Add a flush variant to avoid a potential problem:

If this is a renegotiation, the application data protocol is still running and
adding data to the write buffer. Thus BIO_flush may block until more data than
necessary is written out to network. While it is blocking unnecessarily, the
peer has received the Finished and may send ChangeCipherSpec. The application
data protocol is running at this point, so it may read records and process
ChangeCipherSpec.

I believe this to be the cause of
https://rt.openssl.org/Ticket/Display.html?id=3400 and why it was necessary to
add a CCS_OK in s3_clnt.c. This is worth double-checking, however.

Change-Id: I3eefeb821e7ae53d52dacc587fdc59de9ea9a667