Update patch set 3
Patch Set 3:

(2 comments)

You're suggesting that, during a renegotiation, the application reads a CCS record because it's waiting for a BIO_flush to happen in the handshake?

But in ssl3_read_bytes, if SSL_in_init(s) (which is true when in SSL3_ST_SW_FLUSH) then it'll just call into the handshake function again. That's before checking whether it's a CCS.

Patch-set: 3
Label: Code-Review=0