Update patch set 2
Patch Set 2:

> (1 comment)

I think this change is probably on the right path. The mixture of different methods and different OP_NO_* is a mess - like a lot of things. I'm a little worried about changing things too much before we're internal and in Android too. (It already looks like I'll have to undo some of the DES cleanups for Android.)

As for disabling renego in DTLS: that sounds like a very good idea. There are some hits for "DTLS webrtc renegotiation", but I think they are talking about a different renegotiation. I'll write an email to the WebRTC folks and cc you.

Patch-set: 2
Label: Code-Review=0