Update patch set 3
Patch Set 3:

(2 comments)

Adam: I put it in the session, since as I understand the spec the server does not have to serve the list on session resume. I did not serialize it in ssl_asn1.c because I thought I thought that was for persistent data outliving a browser session (like session tickets). The timestamp list does not seem to be that kind of data.

Patch-set: 3
Label: Code-Review=0