Tidy up Channel ID handshake hash processing.
The handshake hash is processed early, so there is no need to special-case the
message type in ssl3_get_message.

Note that this isn't quite safe a priori. Should we ever read ahead and then
reuse_message this message, this would break. But we never do;
EncryptedExtensions is preceeded by either NextProto (eventually to be gone) or
ChangeCipherSpec. The former does not require lookahead and the latter cannot
lookahead without complicating CCS/Handshake synchronization anyway.

In addition, document the additiona function of some scratch space and
introduce a dedicated state for the hash sampling. Conditioning on init_num
isn't quite ideal; if ssl_read_bytes returns 0 bytes with EWOULDBLOCK, we'll
sample the hash an extra time.

Change-Id: I254945c25a11c493d2da6e7bc332ebe1dc1f2d6f