Introduce a skip_hash parameter to ssl_get_message.
This replaces the special-case in ssl3_get_message for Channel ID. Also add
ssl3_hash_current_message to hash the current message, taking TLS vs DTLS
handshake header size into account.

One subtlety with this flag is that a message intended to be processed with
skip_hash cannot follow an optional message (probed with reuse_message, etc.).
There is an assertion to that effect. If needbe, we can loosen it to requiring
that the preceeding optional message also pass skip_hash and then maintain some
state to perform the more accurate assertion, but this is sufficient for now.

Change-Id: If8c87342b291ac041a35885b9b5ee961aee86eab