Compute the Channel ID hash after ssl_get_message.
This avoids needing the save the hash on the SSL* (and use some field for two
pruposes). Instead, use the new skip_hash option (which actually was already
used here, but at the time, pointlessly). Also fix a minor bug where the hash
would be recomputed in non-blocking mode because init_num may stay zero for a
few state machine iterations.

Change-Id: I3d8331cf3134c5f9a3eda9e988bba5bcebe40933