Refactor server-side CertificateVerify handling.
This moves CertificateVerify digest processing to the new skip_hash flag. It
also refactors it similarly to ssl3_send_cert_verify and moves that logic to a
common ssl3_cert_verify_hash function to compute the handshake hash.

This removes a large chunk of duplicate (and divergent!) logic between TLS and
DTLS. It also removes TLS1_FLAGS_KEEP_HANDSHAKE.

Change-Id: Ia63c94f7d76d901bc9c4c33454fbfede411adf63