Update patch set 3
Patch Set 3:

> No helpful functions to validate received OCSP response?

This is sufficient for Chromium's needs for now. Currently just using it to pull out an SCT extension for Certificate Transparency, which we do externally (this is the same API that NSS exposes).

https://codereview.chromium.org/519473002/

For actual OCSP stapling, we want the response as a DER-encoded blob anyway. On Windows, it gets fed into the platform verifier which doesn't use the same types and wants an opaque blob. Android/Mac's platform verifiers don't have a way to stuff this in. Not sure what CrOS and Linux certificate stories will be, but that's a ways down the road. We can always change this API later, so I just did the simplest thing that worked.

(Not that OCSP stapling itself is particularly useful anyway without perhaps some future must-staple bit...)

Patch-set: 3
Reviewer: Gerrit User 5075 <5075@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=0