Update patch set 1
Patch Set 1:

Actually, given how strict NSS is seems to be, it looks like it's not actually compatible to resume a session from an older version and negotiate a newer one. I'll double-check that and, if so, revise the test and the commit on top of it. (Making it strict means we preserve the invariant in OpenSSL that a resumed SSL_SESSION always has matching s->version and s->session->ssl_version. Dunno if anything cares.)

https://code.google.com/p/chromium/codesearch#chromium/src/net/third_party/nss/ssl/ssl3con.c&q=sid_match&sq=package:chromium&type=cs&l=6576

Though that's puzzling because I see no check in OpenSSL's server code that ensures it never does this, which means such handshakes would fail in Chrome and Firefox...

Patch-set: 1
Reviewer: Gerrit User 5075 <5075@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=0