Update patch set 1
Patch Set 1:

This is rather annoying, but OpenSSL does it too. I don't think we can avoid including both OpenSSL and Windows headers in Chromium. (Some code is going to need to understand both OSCertHandle and X509*)

This does leave us sensitive to wincrypt/boringssl include order, which is a nuisance. Not sure whether all in base.h X509_NAME in base.h and the others in x509.h is better. Or adding #undef blocks in each of them for extra fun.

This seems to get as far as net_unittests (and crypto_unittests doesn't care).

Patch-set: 1