Update patch set 1
Patch Set 1:

(1 comment)

Added a new version of the zero copy API (which was removed when forking OpenSSL). One major difference is that the buffers now can be allocated externally giving caller better control. The API does not add complexity to the caller code when used on asynchronous sockets.

Patch-set: 1
Label: Code-Review=0