Update patch set 3
Patch Set 3:

Do we perhaps want something not based on BIO pairs? That's what we currently use within Chromium and, as I understand, this'll eliminate the copy between Chromium's transport socket and the BIO pair. But there'll still be a copy between the BIO pair and OpenSSL's internal state. The BIO pair itself causes some additional troubles, largely around plumbing of write errors[*]. I'd been wondering if it would be an improvement if we moved away from the BIO pair altogether.


I haven't had the time to actually flesh this out or see how well this actually works with the record layer. (I'm half copying this from some notes I started a bit ago and half typing it out on the spot.) But say we had functions like:

bio_async_result_t BIO_async_read(BIO *bio, size_t *out_bytes_read,
                                  const uint8_t *in, size_t in_len);
bio_async_result_t BIO_async_read_complete(BIO *bio,
                                           size_t *out_bytes_read);
bio_async_result_t BIO_async_write(BIO *bio, size_t *out_bytes_written,
                                   uint8_t *out, size_t out_len);
bio_async_result_t BIO_async_write_complete(BIO *bio,
                                            size_t *out_bytes_written);

where bio_async_result_t is one of:
  BIO_ASYNC_SUCCESS
  BIO_ASYNC_ERROR
  BIO_ASYNC_PENDING

The semantics being, you call BIO_async_read (or write) to indicate that you are okay with ceding control of your buffer asynchronously. If you get BIO_ASYNC_SUCCESS/BIO_ASYNC_ERROR, it's as if BIO_read returned synchronously with success or error. So this API can transparently work with non-async BIOs, BIO_ASYNC_ERROR + should_retry is a valid return code.

If you get BIO_ASYNC_PENDING, that means the operation is in progress and the BIO is holding on to your buffer. When select() or whatever tells you to pump the state machine, you call BIO_async_read_complete which again returns one of the three values. PENDING means it's still not done. SUCCESS/ERROR means the operation has completed and you regain control of your buffer.

This way the SSL state machine can pass in pointers to its record-layer buffer and that buffer transfers all the way through to Chromium's socket implementations.

This also fixes the write error problem because now the SSL state machine has a way to get the asynchronous write error out.

There's a slight awkwardness with wanting to be able to cancel a pending operation. I believe Chromium also lets you regain the buffer by explicitly disconnecting the socket or destroying it to cancel all pending operations. We probably want the explicit disconnect operation for BIO since they're ref-counted so a consumer may not know when they're destroying it.

We'd also be more sensitive to the write patterns OpenSSL makes. With the BIO pair, if OpenSSL calls BIO_write one byte at a time, we have a buffer to guard against that. Now we need OpenSSL to buffer writes reasonably.

SSL_MODE_RELEASE_BUFFERS probably also becomes sillier. As long as there's a hanging BIO_ASYNC_PENDING read, the read buffer needs to stay around. The fact that an SSL implementation needs to maintain a 16k buffer to receive large records is the main reason I'm not so sure about this plan.

Perhaps what we want is to still zero-copy into the record-layer buffer, but initially size it for some smaller buffer size. And if the record turns out to be large, resize (and copy; or allocate a sequence of the smaller buffers as-needed) it for just that record.


I could also imagine tweaking SSL_read/SSL_write in a similar way. Though SSL inherently involves a copy to encrypt/decrypt, so it might not be as useful? It could perhaps avoid the API awkwardness where SSL_write assumes that consecutive calls are consistent because, unlike the UNIX-style write, an SSL implementation can commit to data on the transport before the write has completed. But unlike a generic BIO, it doesn't have to take ownership of the BIO across the asynchronous operation because it has to copy anyway.

This part of my notes is still a big TODO, so past me apparently hadn't thought about it very hard yet. :-)


Thoughts? Adam, how does this mesh with internal consumers' IO model?


[*] Because BIO_write on the BIO pair returns immediately, transport write failures propagate through under a delay in Chromium. If SSLClientSocket::Write fails with a transport write error, it actually means the previous SSLClientSocket::Write failed, which is pretty awkward. It's especially awkward around the last Write an SSLClientSocket ever sees. We actually route that around to appear in Read to avoid the error getting lost. Which is even more awkward as it means Read may flicker back and forth between reporting an error and data.

Patch-set: 3