Update patch set 1
Patch Set 1:

It's starting to get a little late here, but I throw in a couple of questions just to avoid the 24 hour RTT.

Could you give a rough overview of how this could look like from application side? (as in SSLClientSocketOpenSSL)

If I understand you correctly, you want to take the in buffer given by application and write the record protocol data + encrypted data 
directly into the buffer going to the socket?

Patch-set: 1