Fix ECDHE_PSK key exchange.
The current implementation switches the order of other_secret and psk;
other_secret is first. Fix it and rewrite with CBB instead.

DONOTLANDYET: Actually the initial commit flipped it. Some refactoring
accidentally switched it to the correct order without noticing. We may need to
declare that cipher suite cafe always flips it since it's non-standard anyway.
IANA should allocate a different code point for the real one.

Change-Id: Ib52a756aadd66e4bf22c66794447f71f4772da09